Move the `StageBuilder` class into a separate file
