+++ Sources/NativeBuilder/ContainerBuildIR/Graph/GraphBuilder.swift
@@ -452,73 +452,6 @@ public final class GraphBuilder {
     }
 }
 
-/// Builder for individual stages.
-private final class StageBuilder {
-    let name: String?
-    let base: ImageOperation
-    let platform: Platform?
-    private var nodes: [BuildNode] = []
-    private var lastNodeId: UUID?
-    private let analyzers: [any StageAnalyzer]
-    private let reporter: Reporter?
-
-    init(name: String?, base: ImageOperation, platform: Platform? = nil, analyzers: [any StageAnalyzer] = [], reporter: Reporter? = nil) {
-        self.name = name
-        self.base = base
-        self.platform = platform
-        self.analyzers = analyzers
-        self.reporter = reporter
-    }
-
-    @discardableResult
-    func add(_ operation: any Operation, dependsOn: Set<UUID> = []) -> UUID {
-        let node = BuildNode(
-            operation: operation,
-            dependencies: dependsOn
-        )
-
-        // Report node addition
-        if let reporter = reporter {
-            let stageId = name ?? "stage-\(node.id.uuidString.prefix(8))"
-            Task {
-                await reporter.report(
-                    .irEvent(
-                        context: ReportContext(
-                            nodeId: node.id,
-                            stageId: stageId,
-                            description: "Added \(type(of: operation))",
-                            sourceMap: nil
-                        ),
-                        type: .nodeAdded
-                    ))
-            }
-        }
-
-        nodes.append(node)
-        lastNodeId = node.id
-        return node.id
-    }
-
-    func build() throws -> BuildStage {
-        var stage = BuildStage(
-            name: name,
-            base: base,
-            nodes: nodes,
-            platform: platform
-        )
-
-        // Create analysis context
-        let analysisContext = AnalysisContext(reporter: reporter)
-
-        // Run stage analyzers
-        for analyzer in analyzers {
-            stage = try analyzer.analyze(stage, context: analysisContext)
-        }
-
-        return stage
-    }
-}
-
 // MARK: - Convenience Extensions
 
 extension GraphBuilder {
